Time: 4 ms (23.65%) | Memory: 8.9 MB (6.93%) - LeetSync

+++ 171-excel-sheet-column-number/excel-sheet-column-number.cpp
@@ -8,7 +8,7 @@ public:
     }(); 
 
      int titleToNumber(string columnTitle) {
-        long long ans = 0;  // Use long long to avoid overflow
+        int ans = 0;  // Use long long to avoid overflow
         
         for (char c : columnTitle) {
             ans = ans * 26 + (c - 'A' + 1);
